requirements.txt
openai
pyttsx3
SpeechRecognition
pyaudio
pydub
elevenlabs
python-box==6.0.2
pyYAML

-e .